Move site details endpoint to site_evaluation router
